Add files via upload
This project fills your screen with beautiful stars.
It uses Python's turtle module to draw star shapes.
Random numbers scatter the stars over the screen and vary their color, size and shape.

WHAT HAPPENS:
First a nighttime sky is drawn.
Then a single sky appears in the sky.
As the program continues, the sky begins to fill with more and more stars in a wide range of different styles.
The longer you leave the program running the more fantastic and colorful the sky becomes.

The project draws stars one by one but because it uses an infinite while loop it will draw stars forever.
You can change the size, range of the stars by adjusting the limits on the random numbers in the code.


HOW IT WORKS.
The project draws star shapes at random locations in a Turtle Graphics window.
You write a function that can draw a single star. Then you make a loop that repeats it over and over, drawing lots of different stars all over the screen.