Add files via upload
WHAT HAPPENS.
The program will ask you to choose the length and thickness of the line that the turtle paints.
The turtle the scurries around the screen until you stop the program, painting colored lines as it goes.
The type of pattern it makes will change depending on the length and thickness of the lines.

NEXT COLOR:
The choice() function from Python's random module picks a color when you tell turtle to draw a line. 
This means you can't really predict which color the turtle will use each time.

DISPLAY OF COLORS:
Because this program uses an infinite while loop, the turtl keeps drawing until the window is closed
You can not only change the color, width and length.

HOW IT WORKS:
Every pattern in this project is different because the program tells the turtle to face a ransom new direction before painting each line.
The color for each line is also chosen at random from a list of possible colors you've coded.
So  you  can never predict exactly what the turtle will do

RUNAWAY TURTLE:
Given the complete freedom to roam, the turtle tends to wander out of the window.
As you put the program together you'll write some code to check on the turtle's position and stop it from straying too far.
Otherwise, this will turn into a vanishing turtle project.